chore: process review comments
