+++ sdk/src/contracts/collections/features/index.ts
@@ -15,6 +15,7 @@ export * from './getClaimData';
 export * from './getClaimPauseStatus';
 export * from './getDefaultRoyaltyInfo';
 export * from './getLargestTokenId';
+export * from './getOperatorRestriction';
 export * from './getPlatformFees';
 export * from './getPrimarySaleRecipient';
 export * from './getRoyaltyInfoForToken';
@@ -33,7 +34,6 @@ export * from './issueWithinPhaseWithTokenUri';
 export * from './lazyMint';
 export * from './multicall';
 export * from './name';
-export * from './operatorRestriction';
 export * from './owner';
 export * from './ownerOf';
 export * from './royaltyInfo';
